+++ libcrux-ml-kem/src/lib.rs
@@ -19,10 +19,10 @@ mod hash_functions;
 mod ind_cpa;
 mod matrix;
 mod ntt;
-mod polynomial;
+// mod polynomial;
 mod sampling;
 mod serialize;
-mod simd;
+// mod simd;
 mod types;
 
 // Variants
