+++ libcrux-ml-kem/src/ntt.rs
@@ -1,12 +1,10 @@
 use crate::hax_utils::hax_debug_assert;
 
-use crate::{
-    constants::COEFFICIENTS_IN_RING_ELEMENT,
-    polynomial::{
-        invert_ntt_at_layer_1, invert_ntt_at_layer_2, invert_ntt_at_layer_3_plus, ntt_at_layer_1,
-        ntt_at_layer_2, ntt_at_layer_3_plus, ntt_at_layer_7, poly_barrett_reduce,
-        PolynomialRingElement,
-    },
+use crate::constants::COEFFICIENTS_IN_RING_ELEMENT;
+use libcrux_polynomials_avx2::{
+    invert_ntt_at_layer_1, invert_ntt_at_layer_2, invert_ntt_at_layer_3_plus, ntt_at_layer_1,
+    ntt_at_layer_2, ntt_at_layer_3_plus, ntt_at_layer_7, poly_barrett_reduce,
+    PolynomialRingElement,
 };
 
 /// Use the Cooley–Tukey butterfly to compute an in-place NTT representation
