+++ libcrux-ml-kem/src/serialize.rs
@@ -2,10 +2,9 @@ use crate::{
     constants::{BYTES_PER_RING_ELEMENT, SHARED_SECRET_SIZE},
     hax_utils::hax_debug_assert,
     helper::cloop,
-    polynomial::{PolynomialRingElement, VECTORS_IN_RING_ELEMENT},
-    simd,
 };
 use libcrux_polynomials_avx2::traits::{GenericOperations, Operations};
+use libcrux_polynomials_avx2::{vector, PolynomialRingElement, VECTORS_IN_RING_ELEMENT};
 
 #[cfg(hax)]
 use super::constants::COEFFICIENTS_IN_RING_ELEMENT;
@@ -16,9 +15,9 @@ pub(super) fn compress_then_serialize_message(
 ) -> [u8; SHARED_SECRET_SIZE] {
     let mut serialized = [0u8; SHARED_SECRET_SIZE];
     for i in 0..32 {
-        let coefficient = simd::Vector::to_unsigned_representative(re.coefficients[i]);
-        let coefficient_compressed = simd::Vector::compress_1(coefficient);
-        serialized[i] = simd::Vector::serialize_1(coefficient_compressed);
+        let coefficient = vector::Vector::to_unsigned_representative(re.coefficients[i]);
+        let coefficient_compressed = vector::Vector::compress_1(coefficient);
+        serialized[i] = vector::Vector::serialize_1(coefficient_compressed);
     }
     serialized
 }
@@ -28,8 +27,8 @@ pub(super) fn deserialize_then_decompress_message(
 ) -> PolynomialRingElement {
     let mut re = PolynomialRingElement::ZERO();
     for i in 0..32 {
-        let coefficient_compressed = simd::Vector::deserialize_1(serialized[i]);
-        re.coefficients[i] = simd::Vector::decompress_1(coefficient_compressed);
+        let coefficient_compressed = vector::Vector::deserialize_1(serialized[i]);
+        re.coefficients[i] = vector::Vector::decompress_1(coefficient_compressed);
     }
     re
 }
@@ -40,8 +39,8 @@ pub(super) fn serialize_uncompressed_ring_element(
 ) -> [u8; BYTES_PER_RING_ELEMENT] {
     let mut serialized = [0u8; BYTES_PER_RING_ELEMENT];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::to_unsigned_representative(re.coefficients[i]);
-        let bytes = simd::Vector::serialize_12(coefficient);
+        let coefficient = vector::Vector::to_unsigned_representative(re.coefficients[i]);
+        let bytes = vector::Vector::serialize_12(coefficient);
         serialized[12 * i] = bytes[0];
         serialized[12 * i + 1] = bytes[1];
         serialized[12 * i + 2] = bytes[2];
@@ -66,7 +65,7 @@ pub(super) fn deserialize_to_uncompressed_ring_element(serialized: &[u8]) -> Pol
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(12).enumerate() {
-            re.coefficients[i] = simd::Vector::deserialize_12(&bytes);
+            re.coefficients[i] = vector::Vector::deserialize_12(&bytes);
         }
     }
 
@@ -84,8 +83,8 @@ fn deserialize_to_reduced_ring_element(serialized: &[u8]) -> PolynomialRingEleme
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(12).enumerate() {
-            let coefficient = simd::Vector::deserialize_12(&bytes);
-            re.coefficients[i] = simd::Vector::cond_subtract_3329(coefficient);
+            let coefficient = vector::Vector::deserialize_12(&bytes);
+            re.coefficients[i] = vector::Vector::cond_subtract_3329(coefficient);
         }
     }
     re
@@ -115,10 +114,10 @@ pub(super) fn deserialize_ring_elements_reduced<const PUBLIC_KEY_SIZE: usize, co
 fn compress_then_serialize_10<const OUT_LEN: usize>(re: PolynomialRingElement) -> [u8; OUT_LEN] {
     let mut serialized = [0u8; OUT_LEN];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::compress::<10>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes = simd::Vector::serialize_10(coefficient);
+        let coefficient = vector::Vector::compress::<10>(
+            vector::Vector::to_unsigned_representative(re.coefficients[i]),
+        );
+        let bytes = vector::Vector::serialize_10(coefficient);
         serialized[10 * i] = bytes[0];
         serialized[10 * i + 1] = bytes[1];
         serialized[10 * i + 2] = bytes[2];
@@ -137,10 +136,10 @@ fn compress_then_serialize_10<const OUT_LEN: usize>(re: PolynomialRingElement) -
 fn compress_then_serialize_11<const OUT_LEN: usize>(re: PolynomialRingElement) -> [u8; OUT_LEN] {
     let mut serialized = [0u8; OUT_LEN];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::compress::<11>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes = simd::Vector::serialize_11(coefficient);
+        let coefficient = vector::Vector::compress::<11>(
+            vector::Vector::to_unsigned_representative(re.coefficients[i]),
+        );
+        let bytes = vector::Vector::serialize_11(coefficient);
         serialized[11 * i] = bytes[0];
         serialized[11 * i + 1] = bytes[1];
         serialized[11 * i + 2] = bytes[2];
@@ -176,10 +175,10 @@ pub(super) fn compress_then_serialize_ring_element_u<
 fn compress_then_serialize_4<const OUT_LEN: usize>(re: PolynomialRingElement) -> [u8; OUT_LEN] {
     let mut serialized = [0u8; OUT_LEN];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::compress::<4>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes = simd::Vector::serialize_4(coefficient);
+        let coefficient = vector::Vector::compress::<4>(
+            vector::Vector::to_unsigned_representative(re.coefficients[i]),
+        );
+        let bytes = vector::Vector::serialize_4(coefficient);
         serialized[4 * i] = bytes[0];
         serialized[4 * i + 1] = bytes[1];
         serialized[4 * i + 2] = bytes[2];
@@ -193,10 +192,10 @@ fn compress_then_serialize_5<const OUT_LEN: usize>(re: PolynomialRingElement) ->
     let mut serialized = [0u8; OUT_LEN];
 
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficients = simd::Vector::compress::<5>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes5 = simd::Vector::serialize_5(coefficients);
+        let coefficients = vector::Vector::compress::<5>(
+            vector::Vector::to_unsigned_representative(re.coefficients[i]),
+        );
+        let bytes5 = vector::Vector::serialize_5(coefficients);
         serialized[5 * i] = bytes5[0];
         serialized[5 * i + 1] = bytes5[1];
         serialized[5 * i + 2] = bytes5[2];
@@ -230,8 +229,8 @@ fn deserialize_then_decompress_10(serialized: &[u8]) -> PolynomialRingElement {
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(10).enumerate() {
-            let coefficient = simd::Vector::deserialize_10(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<10>(coefficient);
+            let coefficient = vector::Vector::deserialize_10(&bytes);
+            re.coefficients[i] = vector::Vector::decompress::<10>(coefficient);
         }
     }
     re
@@ -245,8 +244,8 @@ fn deserialize_then_decompress_11(serialized: &[u8]) -> PolynomialRingElement {
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(11).enumerate() {
-            let coefficient = simd::Vector::deserialize_11(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<11>(coefficient);
+            let coefficient = vector::Vector::deserialize_11(&bytes);
+            re.coefficients[i] = vector::Vector::decompress::<11>(coefficient);
         }
     }
 
@@ -272,8 +271,8 @@ fn deserialize_then_decompress_4(serialized: &[u8]) -> PolynomialRingElement {
     let mut re = PolynomialRingElement::ZERO();
     cloop! {
         for (i, bytes) in serialized.chunks_exact(4).enumerate() {
-            let coefficient = simd::Vector::deserialize_4(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<4>(coefficient);
+            let coefficient = vector::Vector::deserialize_4(&bytes);
+            re.coefficients[i] = vector::Vector::decompress::<4>(coefficient);
         }
     }
     re
@@ -287,8 +286,8 @@ fn deserialize_then_decompress_5(serialized: &[u8]) -> PolynomialRingElement {
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(5).enumerate() {
-            re.coefficients[i] = simd::Vector::deserialize_5(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<5>(re.coefficients[i]);
+            re.coefficients[i] = vector::Vector::deserialize_5(&bytes);
+            re.coefficients[i] = vector::Vector::decompress::<5>(re.coefficients[i]);
         }
     }
     re
